+++ packages/commercetools/api-client/src/api/createMyOrderFromCart/index.ts
@@ -1,16 +1,19 @@
 import { OrderMyCartCommand } from '../../types/GraphQL';
-import { apolloClient, getSettings } from '../../index';
+import { getCustomQuery, CustomQueryFn, getSettings } from '../../index';
 import CreateMyOrderFromCartMutation from './defaultMutation';
 import { OrderMutationResponse } from '../../types/Api';
+import gql from 'graphql-tag';
 
-const createMyOrderFromCart = async (draft: OrderMyCartCommand): Promise<OrderMutationResponse> => {
-  const { locale, acceptLanguage } = getSettings();
-
-  return await apolloClient.mutate({
-    mutation: CreateMyOrderFromCartMutation,
-    variables: { locale,
-      acceptLanguage,
-      draft },
+const createMyOrderFromCart = async (draft: OrderMyCartCommand, customQueryFn?: CustomQueryFn): Promise<OrderMutationResponse> => {
+  const { locale, acceptLanguage, client } = getSettings();
+  const defaultVariables = { locale,
+    acceptLanguage,
+    draft
+  };
+  const { query, variables } = getCustomQuery(customQueryFn, { defaultQuery: CreateMyOrderFromCartMutation, defaultVariables });
+  return await client.mutate({
+    mutation: gql`${query}`,
+    variables,
     fetchPolicy: 'no-cache'
   });
 };