+++ packages/commercetools/api-client/src/api/getMe/index.ts
@@ -1,7 +1,7 @@
-import { apolloClient, getSettings } from './../../index';
+import { CustomQueryFn, getSettings } from '../../index';
 import { basicProfile, fullProfile } from './defaultQuery';
 import gql from 'graphql-tag';
-import { getCustomQuery } from './../../helpers/queries';
+import { getCustomQuery } from '../../helpers/queries';
 
 interface Options {
   customer?: boolean;
@@ -12,8 +12,9 @@ interface OrdersData {
   me: any;
 }
 
-const getMe = async (params: Options = {}, customQueryFn?) => {
-  const { locale, acceptLanguage } = getSettings();
+const getMe = async (params: Options = {}, customQueryFn?: CustomQueryFn) => {
+  const { locale, acceptLanguage, client } = getSettings();
+
   const { customer }: Options = params;
   const defaultQuery = customer ? fullProfile : basicProfile;
   const defaultVariables = {
@@ -22,7 +23,7 @@ const getMe = async (params: Options = {}, customQueryFn?) => {
   };
   const { query, variables } = getCustomQuery(customQueryFn, { defaultQuery, defaultVariables });
 
-  const request = await apolloClient.query<OrdersData>({
+  const request = await client.query<OrdersData>({
     query: gql`${query}`,
     variables,
     fetchPolicy: 'no-cache'