+++ packages/commercetools/theme/components/Checkout/UserShippingAddresses.vue
@@ -1,11 +1,12 @@
 <template>
   <div>
     <SfAddressPicker
-      :value="currentAddressId"
+      :selected="String(currentAddressId)"
       @input="setCurrentAddress($event)"
-      class="shipping__addresses"
+      class="shipping-addresses"
     >
       <SfAddress
+        class="shipping-addresses__address"
         v-for="shippingAddress in shippingAddresses"
         :key="userShippingGetters.getId(shippingAddress)"
         :name="String(userShippingGetters.getId(shippingAddress))"
@@ -15,7 +16,8 @@
     </SfAddressPicker>
     <SfCheckbox
       data-cy="shipping-details-checkbox_isDefault"
-      v-model="localSetAsDefault"
+      :selected="setAsDefault"
+      @change="$emit('changeSetAsDefault', $event)"
       name="setAsDefault"
       label="Use this address as my default one."
       class="shipping-address-setAsDefault"
@@ -24,10 +26,11 @@
 </template>
 
 <script>
-import { SfCheckbox } from '@storefront-ui/vue';
+import {
+  SfCheckbox,
+  SfAddressPicker
+} from '@storefront-ui/vue';
 import UserShippingAddress from '~/components/UserShippingAddress';
-import SfAddressPicker from '~/components/temp/SfAddressPicker';
-import { ref, watch } from '@vue/composition-api';
 import { userShippingGetters } from '@vue-storefront/commercetools';
 
 export default {
@@ -51,17 +54,11 @@ export default {
     SfAddressPicker,
     UserShippingAddress
   },
-  setup ({ setAsDefault }, { emit }) {
+  setup (_, { emit }) {
     const setCurrentAddress = $event => emit('setCurrentAddress', $event);
-
-    const localSetAsDefault = ref(setAsDefault);
-
-    watch(localSetAsDefault, () => emit('changeSetAsDefault', localSetAsDefault.value));
-    watch(() => setAsDefault, () => localSetAsDefault.value = setAsDefault);
 
     return {
       setCurrentAddress,
-      localSetAsDefault,
       userShippingGetters
     };
   }
@@ -69,21 +66,21 @@ export default {
 </script>
 
 <style lang="scss">
-@import "~@storefront-ui/shared/styles/variables";
+@import "~@storefront-ui/vue/styles";
 
-  .shipping__addresses {
-    @include for-desktop {
-      display: grid;
-      grid-template-columns: 1fr 1fr;
-    }
-    margin-bottom: var(--spacer-xl);
-
-    .sf-address {
-      margin-bottom: var(--spacer-sm);
-    }
+.shipping-addresses {
+  @include for-desktop {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-column-gap: 10px;
   }
-
-  .shipping-address-setAsDefault, .form__action-button--margin-bottom {
-    margin-bottom: var(--spacer-xl);
+  margin-bottom: var(--spacer-xl);
+  &__address {
+    margin-bottom: var(--spacer-sm);
   }
+}
+
+.shipping-address-setAsDefault, .form__action-button--margin-bottom {
+  margin-bottom: var(--spacer-xl);
+}
 </style>