+++ packages/commercetools/theme/pages/Checkout/Payment.vue
@@ -1,10 +1,18 @@
 <template>
   <ValidationObserver v-slot="{ handleSubmit }">
-    <form @submit.prevent="handleSubmit(handleFormSubmit)">
+    <form @submit.prevent="handleSubmit(addressIsModified ? handleBillingAddressSubmit : handleFormSubmit)">
     <SfHeading
       :level="3"
       title="Billing address"
       class="sf-heading--left sf-heading--no-underline title"
+    />
+    <UserBillingAddresses
+      v-if="isAuthenticated && billingAddresses && billingAddresses.length"
+      :setAsDefault="setAsDefault"
+      :billingAddresses="billingAddresses"
+      :currentAddressId="currentAddressId"
+      @setCurrentAddress="setCurrentAddress($event)"
+      @changeSetAsDefault="setAsDefault = $event"
     />
     <SfCheckbox
       :selected="sameAsShipping"
@@ -13,11 +21,11 @@
       name="copyShippingAddress"
       class="form__element"
     />
-      <div class="form">
+      <div class="form" v-if="canAddNewAddress">
         <ValidationProvider name="firstName" rules="required|min:2" v-slot="{ errors }" slim>
           <SfInput
             :value="billingDetails.firstName"
-            @input="firstName => setBillingDetails({ firstName })"
+            @input="firstName => setBillingDetailsAndUnpickAddress({ firstName })"
             label="First name"
             name="firstName"
             class="form__element form__element--half"
@@ -29,7 +37,7 @@
         <ValidationProvider name="lastName" rules="required|min:2" v-slot="{ errors }" slim>
           <SfInput
             :value="billingDetails.lastName"
-            @input="lastName => setBillingDetails({ lastName })"
+            @input="lastName => setBillingDetailsAndUnpickAddress({ lastName })"
             label="Last name"
             name="lastName"
             class="form__element form__element--half form__element--half-even"
@@ -41,7 +49,7 @@
         <ValidationProvider name="streetName" rules="required|min:2" v-slot="{ errors }" slim>
           <SfInput
             :value="billingDetails.streetName"
-            @input="streetName => setBillingDetails({ streetName })"
+            @input="streetName => setBillingDetailsAndUnpickAddress({ streetName })"
             label="Street name"
             name="streetName"
             class="form__element form__element--half"
@@ -53,7 +61,7 @@
         <ValidationProvider name="apartment" rules="required|min:2" v-slot="{ errors }" slim>
           <SfInput
             :value="billingDetails.streetNumber"
-            @input="streetNumber => setBillingDetails({ streetNumber })"
+            @input="streetNumber => setBillingDetailsAndUnpickAddress({ streetNumber })"
             label="House/Apartment number"
             name="apartment"
             class="form__element form__element--half form__element--half-even"
@@ -65,7 +73,7 @@
         <ValidationProvider name="city" rules="required|min:2" v-slot="{ errors }" slim>
           <SfInput
             :value="billingDetails.city"
-            @input="city => setBillingDetails({ city })"
+            @input="city => setBillingDetailsAndUnpickAddress({ city })"
             label="City"
             name="city"
             class="form__element form__element--half"
@@ -77,7 +85,7 @@
         <ValidationProvider name="zipCode" rules="required|min:2" v-slot="{ errors }" slim>
           <SfInput
             :value="billingDetails.postalCode"
-            @input="postalCode => setBillingDetails({ postalCode })"
+            @input="postalCode => setBillingDetailsAndUnpickAddress({ postalCode })"
             label="Zip-code"
             name="zipCode"
             class="form__element form__element--half form__element--half-even"
@@ -89,7 +97,7 @@
         <ValidationProvider name="country" rules="required|min:2" v-slot="{ errors }" slim>
           <SfSelect
             :selected="billingDetails.country"
-            @change="country => setBillingDetails({ country })"
+            @change="country => setBillingDetailsAndUnpickAddress({ country })"
             label="Country"
             class="form__element form__element--half form__select sf-select--underlined"
             required
@@ -108,7 +116,7 @@
         <ValidationProvider name="phone" rules="required|min:2" v-slot="{ errors }" slim>
           <SfInput
             :value="billingDetails.contactInfo.phone"
-            @input="phone => setBillingDetails({ contactInfo: { phone } })"
+            @input="phone => setBillingDetailsAndUnpickAddress({ contactInfo: { phone } })"
             label="Phone number"
             name="phone"
             class="form__element form__element--half form__element--half-even"
@@ -118,13 +126,22 @@
           />
         </ValidationProvider>
       </div>
+      <SfButton
+        v-if="!canAddNewAddress"
+        class="form__action-button form__action-button--margin-bottom"
+        type="submit"
+        @click.native="canAddNewAddress = true"
+      >
+        Add new address
+      </SfButton>
     <SfHeading
+      v-if="canContinueToReview"
       :level="3"
       title="Payment methods"
       class="sf-heading--left sf-heading--no-underline title"
     />
     <div class="form">
-      <div class="form__element payment-methods">
+      <div class="form__element payment-methods" v-if="canContinueToReview">
         <SfRadio
           v-for="item in paymentMethods"
           :key="item.value"
@@ -145,8 +162,11 @@
       </div>
       <div class="form__action">
         <nuxt-link to="/checkout/shipping" class="sf-button color-secondary form__back-button">Go back</nuxt-link>
-        <SfButton class="form__action-button" type="submit" :disabled="loading.billingAddress">
+        <SfButton class="form__action-button" type="submit" :disabled="loading.billingAddress" v-if="canContinueToReview">
           Review my order
+        </SfButton>
+        <SfButton class="form__action-button" type="submit" :disabled="loading.billingAddress" v-else>
+          Select payment method
         </SfButton>
       </div>
     </div>
@@ -164,11 +184,12 @@ import {
   SfImage,
   SfCheckbox
 } from '@storefront-ui/vue';
-import { ref } from '@vue/composition-api';
-import { useCheckout } from '@vue-storefront/commercetools';
+import { ref, computed, onMounted } from '@vue/composition-api';
+import { useCheckout, useUser, userBillingGetters, useUserBilling } from '@vue-storefront/commercetools';
 import { ValidationProvider, ValidationObserver, extend } from 'vee-validate';
 import { required, min } from 'vee-validate/dist/rules';
 import { onSSR } from '@vue-storefront/core';
+
 const COUNTRIES = [
   { key: 'US',
     label: 'United States' },
@@ -197,6 +218,7 @@ export default {
     SfRadio,
     SfImage,
     SfCheckbox,
+    UserBillingAddresses: () => import('~/components/Checkout/UserBillingAddresses'),
     ValidationProvider,
     ValidationObserver
   },
@@ -207,30 +229,107 @@ export default {
       paymentMethods,
       chosenPaymentMethod,
       loadPaymentMethods,
+      isBillingAddressCompleted,
       setBillingDetails,
       setPaymentMethod,
       loadDetails,
       loading
     } = useCheckout();
+    const { billing, load: loadBilling, setDefault } = useUserBilling();
+    const { isAuthenticated } = useUser();
+
+    const canAddNewAddress = ref(true);
+    const addressIsModified = ref(false);
+    const currentAddressId = ref(-1);
+    const setAsDefault = ref(false);
+
     const sameAsShipping = ref(false);
     let oldBilling = null;
+
+    const mapAbstractAddressToIntegrationAddress = address => ({
+      ...billingDetails.value,
+      contactInfo: {
+        ...billingDetails.value.contactInfo,
+        phone: userBillingGetters.getPhone(address)
+      },
+      streetNumber: userBillingGetters.getApartmentNumber(address),
+      city: userBillingGetters.getCity(address),
+      country: userBillingGetters.getCountry(address),
+      state: userBillingGetters.getProvince(address),
+      firstName: userBillingGetters.getFirstName(address),
+      lastName: userBillingGetters.getLastName(address),
+      streetName: userBillingGetters.getStreetName(address),
+      postalCode: userBillingGetters.getPostCode(address)
+    });
+
+    const setCurrentAddress = async (addressId) => {
+      const chosenAddress = userBillingGetters.getAddresses(billing.value, { id: addressId });
+      if (!chosenAddress || !chosenAddress.length) {
+        return;
+      }
+      currentAddressId.value = addressId;
+      setBillingDetails(mapAbstractAddressToIntegrationAddress(chosenAddress[0]));
+      addressIsModified.value = true;
+      sameAsShipping.value = false;
+    };
+
     onSSR(async () => {
       await loadDetails();
       await loadPaymentMethods();
     });
+
+    onMounted(async () => {
+      if (isAuthenticated.value) {
+        await loadBilling();
+        const billingAddresses = userBillingGetters.getAddresses(billing.value);
+        if (!billingAddresses || !billingAddresses.length) {
+          return;
+        }
+        canAddNewAddress.value = false;
+        if (userBillingGetters.isDefault(billingAddresses[0])) {
+          setCurrentAddress(userBillingGetters.getId(billingAddresses[0]));
+        }
+      }
+    });
+
     const handleFormSubmit = async () => {
       await setBillingDetails(billingDetails.value, { save: true });
       context.root.$router.push('/checkout/order-review');
     };
+
+    const handleBillingAddressSubmit = async () => {
+      await setBillingDetails(billingDetails.value, { save: true });
+      // here we should have something like loadPaymentMethods which depends on billing address imo
+      if (currentAddressId.value > -1 && setAsDefault.value) {
+        const chosenAddress = userBillingGetters.getAddresses(billing.value, { id: currentAddressId.value });
+        if (!chosenAddress || !chosenAddress.length) {
+          return;
+        }
+        await setDefault(chosenAddress[0]);
+      }
+      addressIsModified.value = false;
+    };
+
     const handleCheckSameAddress = () => {
       sameAsShipping.value = !sameAsShipping.value;
       if (sameAsShipping.value) {
         oldBilling = billingDetails.value;
         setBillingDetails(shippingDetails.value);
+        currentAddressId.value = -1;
+        setAsDefault.value = false;
+        addressIsModified.value = true;
         return;
       }
       setBillingDetails(oldBilling);
     };
+
+    const setBillingDetailsAndUnpickAddress = value => {
+      setBillingDetails(value);
+      currentAddressId.value = -1;
+      addressIsModified.value = true;
+      sameAsShipping.value = false;
+    };
+
     return {
       loading,
       billingDetails,
@@ -238,10 +337,20 @@ export default {
       chosenPaymentMethod,
       sameAsShipping,
       COUNTRIES,
-      setBillingDetails,
       setPaymentMethod,
       handleFormSubmit,
-      handleCheckSameAddress
+      handleBillingAddressSubmit,
+      handleCheckSameAddress,
+      isAuthenticated,
+      billingAddresses: computed(() => userBillingGetters.getAddresses(billing.value)),
+      setAsDefault,
+      canContinueToReview: computed(() => isBillingAddressCompleted.value && !addressIsModified.value),
+      currentAddressId: computed(() => currentAddressId.value),
+      setCurrentAddress,
+      addressIsModified,
+      setBillingDetailsAndUnpickAddress,
+      canAddNewAddress,
+      userBillingGetters
     };
   }
 };
@@ -320,10 +429,10 @@ export default {
       margin: 0 var(--spacer-xl) 0 0;
     }
   }
-  &__button {
-    --button-width: 100%;
+  &__back-button {
+    margin: 0 0 var(--spacer-sm) 0;
     @include for-desktop {
-      --button-width: auto;
+      margin: 0 var(--spacer-xl) 0 0;
     }
   }
 }
@@ -339,6 +448,7 @@ export default {
   --radio-container-align-items: center;
   --ratio-content-margin: 0 0 0 var(--spacer-base);
   --radio-label-font-size: var(--font-base);
+  --radio-background: transparent;
   white-space: nowrap;
   border: 1px solid var(--c-light);
   border-width: 1px 0 0 0;