+++ packages/core/core/__tests__/factories/useUserBillingFactory.spec.ts
@@ -20,15 +20,11 @@ describe('[CORE - factories] useUserBillingFactory', () => {
   it('should have proper initial properties', () => {
     const { useUserBilling } = useUserBillingFactory(factoryParams);
     const {
-      addresses,
-      totalAddresses,
-      defaultAddress,
+      billing,
       loading
     } = useUserBilling();
 
-    expect(addresses.value).toEqual([]);
-    expect(totalAddresses.value).toEqual(0);
-    expect(defaultAddress.value).toEqual(null);
+    expect(billing.value).toEqual({});
     expect(loading.value).toEqual(false);
   });
 
@@ -38,7 +34,7 @@ describe('[CORE - factories] useUserBillingFactory', () => {
         const paramsToUpdate = { name: 'Test'};
         factoryParams.addAddress.mockReturnValueOnce(paramsToUpdate);
         await useUserBillingMethods.addAddress(paramsToUpdate);
-        expect(useUserBillingMethods.addresses.value).toEqual(paramsToUpdate);
+        expect(useUserBillingMethods.billing.value).toEqual(paramsToUpdate);
       });
 
       it('throws error', async () => {
@@ -58,7 +54,7 @@ describe('[CORE - factories] useUserBillingFactory', () => {
         const paramsToUpdate = { name: 'Test'};
         factoryParams.deleteAddress.mockReturnValueOnce(paramsToUpdate);
         await useUserBillingMethods.deleteAddress(paramsToUpdate);
-        expect(useUserBillingMethods.addresses.value).toEqual(paramsToUpdate);
+        expect(useUserBillingMethods.billing.value).toEqual(paramsToUpdate);
       });
 
       it('throws error', async () => {
@@ -78,7 +74,7 @@ describe('[CORE - factories] useUserBillingFactory', () => {
         const paramsToUpdate = { name: 'Test'};
         factoryParams.updateAddress.mockReturnValueOnce(paramsToUpdate);
         await useUserBillingMethods.updateAddress(paramsToUpdate);
-        expect(useUserBillingMethods.addresses.value).toEqual(paramsToUpdate);
+        expect(useUserBillingMethods.billing.value).toEqual(paramsToUpdate);
       });
 
       it('throws error', async () => {
@@ -98,7 +94,7 @@ describe('[CORE - factories] useUserBillingFactory', () => {
         const paramsToUpdate = { name: 'Test'};
         factoryParams.load.mockReturnValueOnce(paramsToUpdate);
         await useUserBillingMethods.load();
-        expect(useUserBillingMethods.addresses.value).toEqual(paramsToUpdate);
+        expect(useUserBillingMethods.billing.value).toEqual(paramsToUpdate);
       });
 
       it('throws error', async () => {
@@ -118,7 +114,7 @@ describe('[CORE - factories] useUserBillingFactory', () => {
         const paramsToUpdate = { name: 'Test'};
         factoryParams.setDefault.mockReturnValueOnce(paramsToUpdate);
         await useUserBillingMethods.setDefault(paramsToUpdate);
-        expect(useUserBillingMethods.defaultAddress.value).toEqual(paramsToUpdate);
+        expect(useUserBillingMethods.billing.value).toEqual(paramsToUpdate);
       });
 
       it('throws error', async () => {