+++ packages/core/core/__tests__/middleware/createMiddleware.spec.ts
@@ -1,4 +1,3 @@
-
 import { createMiddleware } from '../../src/server';
 
 jest.mock('express', () => ({