+++ packages/core/nuxt-theme-module/theme/pages/Category.vue
@@ -4,7 +4,6 @@
       class="breadcrumbs desktop-only"
       :breadcrumbs="breadcrumbs"
     />
-
     <div class="navbar section">
       <div class="navbar__aside desktop-only">
         <SfHeading :level="3" :title="$t('Categories')" class="navbar__title" />
@@ -253,8 +252,7 @@ import {
 } from '@storefront-ui/vue';
 import { computed, onMounted } from '@vue/composition-api';
 import { useCart, useFacet, useWishlist, facetGetters, productGetters } from '<%= options.generate.replace.composables %>';
-import { useUiHelpers } from '~/composables';
-import uiState from '~/assets/ui-state';
+import { useUiHelpers, useUiState } from '~/composables';
 import { onSSR } from '@vue-storefront/core';
 import Filters from '../components/Filters';
 
@@ -263,6 +261,7 @@ export default {
   setup(props, context) {
     onMounted(() => context.root.$scrollTo(context.root.$el, 2000));
     const th = useUiHelpers();
+    const uiState = useUiState();
     const { addToCart, isOnCart } = useCart();
     const { addToWishlist } = useWishlist();
     const { result, search, loading } = useFacet();